+++ rltk/geometry.cpp
@@ -10,7 +10,7 @@ namespace rltk {
  */
 std::pair<int, int> project_angle(const int x, const int y, const double radius, const double degrees_radians)
 {
-	return std::make_pair(x + radius * std::cos(degrees_radians), y + radius * std::sin(degrees_radians));
+	return std::make_pair(static_cast<int>(x + radius * std::cos(degrees_radians)), static_cast<int>(y + radius * std::sin(degrees_radians)));
 }
 
 /*
@@ -20,12 +20,12 @@ std::pair<int, int> project_angle(const int x, const int y, const double radius,
  */
 void line_func_3d(int x1, int y1, int z1, const int x2, const int y2, const int z2, std::function<void(int, int, int)> &&func)
 {
-	float x = x1;
-	float y = y1;
-	float z = z1;
+	float x = static_cast<float>(x1);
+	float y = static_cast<float>(y1);
+	float z = static_cast<float>(z1);
 
 	float length = distance3d(x1, y1, z1, x2, y2, z2);
-	int steps = std::floor(length);
+	int steps = static_cast<int>(std::floor(length));
 	float x_step = (x - x2) / length;
 	float y_step = (y - y2) / length;
 	float z_step = (z - z2) / length;
@@ -34,7 +34,7 @@ void line_func_3d(int x1, int y1, int z1, const int x2, const int y2, const int 
 		x += x_step;
 		y += y_step;
 		z += z_step;
-		func(std::floor(x), std::floor(y), std::floor(z));
+		func(static_cast<int>(std::floor(x)), static_cast<int>(std::floor(y)), static_cast<int>(std::floor(z)));
 	}
 }
 
@@ -45,12 +45,12 @@ void line_func_3d(int x1, int y1, int z1, const int x2, const int y2, const int 
  */
 void line_func_3d_cancellable(int x1, int y1, int z1, const int x2, const int y2, const int z2, std::function<bool(int, int, int)> &&func)
 {
-	float x = x1;
-	float y = y1;
-	float z = z1;
+	float x = static_cast<float>(x1);
+	float y = static_cast<float>(y1);
+	float z = static_cast<float>(z1);
 
 	float length = distance3d(x1, y1, z1, x2, y2, z2);
-	int steps = std::floor(length);
+	int steps = static_cast<int>(std::floor(length));
 	float x_step = (x - x2) / length;
 	float y_step = (y - y2) / length;
 	float z_step = (z - z2) / length;
@@ -59,7 +59,7 @@ void line_func_3d_cancellable(int x1, int y1, int z1, const int x2, const int y2
 		x += x_step;
 		y += y_step;
 		z += z_step;
-		const bool keep_going = func(std::floor(x), std::floor(y), std::floor(z));
+		const bool keep_going = func(static_cast<int>(std::floor(x)), static_cast<int>(std::floor(y)), static_cast<int>(std::floor(z)));
 		if (!keep_going) return;
 	}
 }
