+++ rltk/rng.cpp
@@ -11,20 +11,20 @@ namespace rltk
 
 random_number_generator::random_number_generator()
 {
-	initial_seed = std::random_device()();
-	rng.seed(initial_seed);
+	initial_seed = static_cast<uint64_t>(std::random_device()());
+	rng.seed(static_cast<unsigned int>(initial_seed));
 }
 
 random_number_generator::random_number_generator(const uint64_t seed)
 {
 	initial_seed = seed;
-	rng.seed(seed);
+	rng.seed(static_cast<unsigned int>(seed));
 }
 
 random_number_generator::random_number_generator(const std::string seed) {
 	std::hash<std::string> hash_func;
 	initial_seed = hash_func(seed);
-	rng.seed(initial_seed);
+	rng.seed(static_cast<unsigned int>(initial_seed));
 }
 
 int random_number_generator::roll_dice(const int& n, const int& d)
